requirements.txt
aiohttp==3.11.12
alembic==1.14.1
annotated-types==0.7.0
anyio==4.8.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asynclog==0.1.7
asyncpg==0.30.0
async_factory_boy==1.0.1
bcrypt==4.2.1
certifi==2025.1.31
cffi==1.17.1
cfgv==3.4.0
click==8.1.8
colorama==0.4.6
cryptography==44.0.0
distlib==0.3.9
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.8
fastapi-cache2==0.2.2
fastapi-users==14.0.1
fastapi-users-db-sqlalchemy==7.0.0
filelock==3.17.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.27.2
identify==2.6.7
idna==3.10
iniconfig==2.0.0
makefun==1.15.6
Mako==1.3.9
MarkupSafe==3.0.2
mypy-extensions==1.0.0
nodeenv==1.9.1
packaging==24.2
pathspec==0.12.1
pendulum==3.0.0
platformdirs==4.3.6
pluggy==1.5.0
pwdlib==0.2.1
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.6.1
pydantic_core==2.27.2
PyJWT==2.10.1
pytest==8.3.4
pytest-asyncio==0.24.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
PyYAML==6.0.2
redis==4.6.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.38
starlette==0.45.3
time-machine==2.16.0
typing_extensions==4.12.2
tzdata==2025.1
uvicorn==0.34.0
virtualenv==20.29.1
watchfiles==1.0.4
websockets==14.2
